Little update on roaming
- Updated script and enemy prefab on roaming pause time.

+++ Earnest-Dinosaurs Prototype/Assets/Scripts/enemyAI.cs
@@ -20,7 +20,7 @@ public class enemyAI : MonoBehaviour, IDamage
     [SerializeField] int viewCone;
     [SerializeField] int shootCone;
     [SerializeField] int roamDist;
-    [SerializeField] int roamPauseTime;
+    [SerializeField] float roamPauseTime;
 
     [Header("----- Enemy gun's Stats ------")]
     [SerializeField] GameObject bulletObject;
